Moved public static buildAnnotation to J.Annotation
